requirements.txt
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
google-auth==2.34.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
h11==0.14.0
httplib2==0.22.0
idna==3.8
oauth2client==4.1.3
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyparsing==3.1.4
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.24.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
webdriver-manager==4.0.2
websocket-client==1.8.0
wget==3.2
wsproto==1.2.0
